% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrir_livro.R
\name{abrir_livro}
\alias{abrir_livro}
\title{Abrir livro da oficina}
\usage{
abrir_livro()
}
\value{
O livro é aberto no navegador
}
\description{
Abre o livro que contém os comandos da oficina
}
\examples{
abrir_livro()

}
\author{
Alexandre Jaloto
}
