% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco2pl.t1.R
\docType{data}
\name{banco2pl.t1}
\alias{banco2pl.t1}
\title{Banco simulado de itens dicotômicos}
\format{
An object of class \code{data.frame} with 1000 rows and 50 columns.
}
\usage{
banco2pl.t1
}
\description{
Banco simulado de itens dicotômicos
}
\details{
Banco simulado com 500 sujeitos e 10 itens dicotômicos, em que 1
equivale à resposta correta e 0, à errada. Aplicação inicial do instrumento
com 10 itens.
}
\keyword{datasets}
