% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco2pl.t2.R
\docType{data}
\name{banco2pl.t2}
\alias{banco2pl.t2}
\title{Banco simulado de itens dicotômicos}
\format{
An object of class \code{data.frame} with 1000 rows and 50 columns.
}
\usage{
banco2pl.t2
}
\description{
Banco simulado de itens dicotômicos
}
\details{
Banco simulado com 500 sujeitos e 10 itens dicotômicos, em que 1
equivale à resposta correta e 0, à errada. Segunda aplicação inicial do
instrumento com 10 itens, sendo cinco comuns à primeira aplicação
(\code{banco.2pl.t2}).
}
\keyword{datasets}
