% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.rmsd.R
\name{dif.rmsd}
\alias{dif.rmsd}
\title{RMSD}
\usage{
dif.rmsd(calib, flag = 0, probfun = TRUE, dentype = "norm")
}
\arguments{
\item{calib}{objeto do tipo \code{SingleGroupClass} gerado com a função \code{mirt::mirt}}

\item{flag}{ponto de corte para sinalizar itens com DIF}

\item{probfun}{argumento lógico. Se FALSE, o função de escore esperado
será integrada}

\item{dentype}{densidade para usar para o traço latente. Pode ser \code{'norm'}
para densidade normal com média e variância extraídas do objeto,
\code{'snorm'} para normal padrão e \code{'empirical'} para usar a densidade
estimada obtida no passo E}
}
\description{
Análise de DIF usando Root Mean Square Deviaton (RMSD)
}
\examples{
data(banco2pl.t1, package = 'oficinaTRI2025')
fit.t1 <- mirt(data = banco2pl.t1, model = 1, itemtype = '2PL', TOL = .001)
data(banco2pl.t2, package = 'oficinaTRI2025')
tab.pars <- mirt(data = banco2pl.t2, model = 1, itemtype = '2PL', TOL = .001, pars = 'values')
itens.comuns <- paste0('Item_', 21:50)
pars.comuns <- mod2values(fit.t1)
for(i in 1:length(itens.comuns))
{
  tab.pars[tab.pars$item == itens.comuns[i] & tab.pars$name == 'a1', 'value'] <- pars.comuns[pars.comuns$item == itens.comuns[i] & pars.comuns$name == 'a1', 'value']
  tab.pars[tab.pars$item == itens.comuns[i] & tab.pars$name == 'd', 'value'] <- pars.comuns[pars.comuns$item == itens.comuns[i] & pars.comuns$name == 'd', 'value']
  tab.pars[tab.pars$item == itens.comuns[i], 'est'] <- FALSE
}
tab.pars[tab.pars$name == 'MEAN_1', 'est'] <- TRUE
tab.pars[tab.pars$name == 'COV_11', 'est'] <- TRUE
fit.t2 <- mirt(data = banco2pl.t2, model = 1, itemtype = '2PL', TOL = .001, pars = tab.pars)
rmsd.pisa(calib = fit.t2, flag = .1)

}
