% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste.R
\name{teste}
\alias{teste}
\title{Testar pacote}
\usage{
teste()
}
\value{
Caso o teste não tenha identificado erro, uma
mensagem será emitida no console e o livro do curso será
aberto no navegador.
}
\description{
Testa se o pacote foi instalado corretamente,
bem como as dependências dos pacotes.
}
\examples{

teste()
}
\author{
Alexandre Jaloto
}
